pyproject.toml
[tool.poetry]
name = "sales-hack"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
setuptools = "^74.1.2"
sqlalchemy = "^2.0.32"
python-dotenv = "^1.0.1"
asyncpg = "^0.29.0"
pydantic = "^2.8.2"
fastapi = {extras = ["standart"], version = "^0.112.2"}
uvicorn = "^0.30.6"
streamlit = "^1.38.0"
openpyxl = "^3.1.5"
tqdm = "^4.66.5"
pgvector = "^0.3.2"
clickhouse-driver = "^0.2.9"
typing-extensions = "^4.12.2"
psycopg2-binary = "^2.9.9"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
